requirements.txt
opencv-python
numpy
pyttsx3
ultralytics
python-dotenv
twilio
requests